+++ testing/TestFunc.cpp
@@ -7,6 +7,7 @@
 #include <VaLib/Types/String.hpp>
 
 #include <VaLib/FuncTools/Func.hpp>
+#include <VaLib/FuncTools/TypeWrapper.hpp>
 #include <VaLib/FuncTools/Partial.hpp>
 
 // test cases
@@ -14,6 +15,7 @@ int add2(int a, int b) { return a + b; }
 int add3(int a, int b, int c) { return a + b + c; }
 int add4(int a, int b, int c, int d) { return a + b + c + d; }
 int times2(int& val) { return val * 2; }
+void noReturnFunction() {}
 
 int return123() {
     return 123;
@@ -46,6 +48,23 @@ bool testPartial(testing::Test& t) {
     return t.success();
 }
 
+// VaTypeWrapper works only on C++17+
+#if __cplusplus >= CPP17
+bool testTypeWrapper(testing::Test& t) {
+    using AddWrapper = VaTypeWrapper<add2>;
+    if (AddWrapper{}(10, 20) != 30) {
+        return t.fail("unexpected result");
+    }
+
+    using VoidFuncWrapper = VaTypeWrapper<noReturnFunction>;
+    VoidFuncWrapper{}();
+
+    return t.success();
+}
+#else
+bool testTypeWrapper(testing::Test&) {}
+#endif
+
 bool testFunc(testing::Test& t) {
     VaFunc<int(int, int)> add = [](int a, int b) { return a + b; };
     if (add(2, 3) != 5) {
@@ -62,7 +81,11 @@ bool testFunc(testing::Test& t) {
         return t.fail("unexpected result (other callable class)");
     }
 
+    VaFunc<void()> fn3 = noReturnFunction;
+    fn3();
+
     if (!t.helper(testPartial)) return false;
+    if (!t.helper(testTypeWrapper)) return false;
     return t.success();
 }