+++ CHANGELOG.md
@@ -14,3 +14,4 @@
 - Improved the VaAny and VaFunc classes to no longer use the C++23-deprecated std::aligned_storage.
 ### Added
 - Added a new `VaMethod` class for storing class methods. Minor fixes in `VaFunc`.
+- Added VaTypeWrapper that wraps a function into a struct with operator().